+++ integrations/azure-devops/CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
-- Added folder webhook processor
+- Added live events for folder kind
 
 
 ## 0.2.2 (2025-05-27)