+++ integrations/azure-devops/CHANGELOG.md
@@ -7,8 +7,69 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
-## 0.1.153 (2025-05-20)
+## 0.2.4 (2025-05-28)
 
+
+### Improvements
+
+- Bumped ocean version to ^0.23.3
+
+
+## 0.2.3 (2025-05-28)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.23.2
+
+
+## 0.2.2 (2025-05-27)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.23.1
+
+
+## 0.2.1 (2025-05-27)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.23.0
+
+
+## 0.2.0 (2025-05-26)
+
+### Features
+
+ Transitioned live events management to ocean’s `LiveEventProcessorManager` to streamline processing
+
+## 0.1.156 (2025-05-26)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.22.12
+
+
+## 0.1.155 (2025-05-26)
+
+
+### Improvements
+
+- Bumped ocean version to ^0.22.11
+
+
+## 0.1.154 (2025-05-21)
+
+
+### Improvements
+
+- Added support for monorepo
+
+
+## 0.1.153 (2025-05-20)
 
 ### Improvements
 
@@ -56,6 +117,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 ## 0.1.147 (2025-04-27)
+
 
 ### Bug Fixes